// src/app/interfaces/learning-need.ts
export interface LearningNeedSectionOne {
  id?: string;
  author?: string;
  jobTitle?: string;
  quote?: string;
  para1?: string;
  para2?: string;
}

export interface LearningNeedSectionTwo {
  id?: string;
  title?: string;
  titleSpan?: string;
  desc?: string;
  btnText?: string;
  btnLink?: string;
}

export interface LdProcesses {
  id?: string;
  title?: string;
  position?: any;
  desc?: string;
  para?: string;
}